+++ ts/src/coder/borsh/state.ts
@@ -1,9 +1,9 @@
 import { Buffer } from "buffer";
 import { Layout } from "buffer-layout";
 import { sha256 } from "js-sha256";
-import { Idl } from "../idl.js";
+import { Idl } from "../../idl.js";
 import { IdlCoder } from "./idl.js";
-import * as features from "../utils/features.js";
+import * as features from "../../utils/features.js";
 
 export class BorshStateCoder {
   private layout: Layout;
