+++ ts/src/coder/index.ts
@@ -1,16 +1,7 @@
-import { Idl, IdlEvent } from "../idl.js";
-import { BorshInstructionCoder } from "./instruction.js";
-import { BorshAccountsCoder } from "./accounts.js";
-import { BorshEventCoder } from "./event.js";
-import { BorshStateCoder } from "./state.js";
-import { sighash } from "./common.js";
+import { IdlEvent } from "../idl.js";
 import { Event } from "../program/event.js";
 
-export { accountSize } from "./common.js";
-export { BorshInstructionCoder } from "./instruction.js";
-export { BorshAccountsCoder, ACCOUNT_DISCRIMINATOR_SIZE } from "./accounts.js";
-export { BorshEventCoder, eventDiscriminator } from "./event.js";
-export { BorshStateCoder, stateDiscriminator } from "./state.js";
+export * from "./borsh/index.js";
 
 /**
  * Coder provides a facade for encoding and decoding all IDL related objects.
@@ -58,42 +49,4 @@ export interface EventCoder {
   decode<E extends IdlEvent = IdlEvent, T = Record<string, string>>(
     log: string
   ): Event<E, T> | null;
-}
-
-/**
- * BorshCoder is the default Coder for Anchor programs.
- */
-export class BorshCoder<A extends string = string> implements Coder {
-  /**
-   * Instruction coder.
-   */
-  readonly instruction: BorshInstructionCoder;
-
-  /**
-   * Account coder.
-   */
-  readonly accounts: BorshAccountsCoder<A>;
-
-  /**
-   * Coder for state structs.
-   */
-  readonly state: BorshStateCoder;
-
-  /**
-   * Coder for events.
-   */
-  readonly events: BorshEventCoder;
-
-  constructor(idl: Idl) {
-    this.instruction = new BorshInstructionCoder(idl);
-    this.accounts = new BorshAccountsCoder(idl);
-    this.events = new BorshEventCoder(idl);
-    if (idl.state) {
-      this.state = new BorshStateCoder(idl);
-    }
-  }
-
-  public sighash(nameSpace: string, ixName: string): Buffer {
-    return sighash(nameSpace, ixName);
-  }
 }
