+++ ts/src/index.ts
@@ -3,21 +3,10 @@ import { isBrowser } from "./utils/common.js";
 export { default as BN } from "bn.js";
 export * as web3 from "@solana/web3.js";
 export { default as Provider, getProvider, setProvider } from "./provider.js";
-export {
-  Coder,
-  InstructionCoder,
-  EventCoder,
-  StateCoder,
-  AccountsCoder,
-  BorshCoder,
-  BorshInstructionCoder,
-  BorshEventCoder,
-  BorshStateCoder,
-  BorshAccountsCoder,
-} from "./coder/index.js";
 export * from "./error.js";
-export { Instruction } from "./coder/instruction.js";
+export { Instruction } from "./coder/borsh/instruction.js";
 export { Idl } from "./idl.js";
+export * from "./coder/index.js";
 export * as utils from "./utils/index.js";
 export * from "./program/index.js";
